Added package DiceRandomizer
A Java program that prompts for the number of dice and the type of dice
the user would like to roll. Currently the program is just built to take
in one type of dice. Below are the goals I have for the application.
Please note some of these goals involve other project ideas I have.

Next task is to improve upon the (Pseudo) Random number generator
May want to look at timing/efficiency of the program - Not sure how to
do this yet
May want to consider a GUI or some other interface - I think I may just
keep a text interface
Another possible task is to implement this in a situation of inheritance
for a DND Character sheet